Yet more problems fixed in the alpha port.   -- more int -> long checking.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@221 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/extensible/struct.h
@@ -11,7 +11,7 @@ struct extensible
    int result;
    char output[STRMAX];
    struct extensible *next;
-   int miboid[MIBMAX];
+   unsigned long miboid[MIBMAX];
    int miblen;
    int pid;
 };
